backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiofiles==23.2.1
aiosqlite==0.19.0
httpx==0.25.1
gitpython==3.1.40
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
certifi>=2023.7.22